third_party/WebKit/LayoutTests/http/tests/feature-policy-vibrate/vibrate-allowed-by-container-policy-relocate-and-no-reload-expected.txt

CONSOLE WARNING: line 4: Unrecognized feature: 'vibrate'.
This is a testharness.js-based test.
FAIL Iframe navigated itself to a new origin, vibrate is disabled by container policy. assert_false: navigator.vibrate(): expected false got true
Harness: the test ran to completion.
